+++ src/renderer/components/UI/VnasPanel.tsx
@@ -196,6 +196,7 @@ export function VnasPanel() {
                     <option value="live">Live</option>
                     <option value="sweatbox1">Sweatbox 1</option>
                     <option value="sweatbox2">Sweatbox 2</option>
+                    <option value="test">Test</option>
                   </select>
                 </div>